server/requirement.txt
fastapi
uvicorn
prisma
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
python-multipart